veille_osint/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
APScheduler==3.11.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
black==24.10.0
boto3==1.40.1
botocore==1.40.1
cachetools==5.5.2
celery==5.5.3
certego_saas==0.7.11
certifi==2025.7.14
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.1
distlib==0.4.0
Django==5.2.4
django-cache-memoize==0.2.1
django-email-utils==1.0.0
django-filter==25.1
django-ipware==7.0.1
django-rest-durin==1.0.0
django-user-agents==0.4.0
djangorestframework==3.16.0
djangorestframework-filters==1.0.0.dev2
drf-flex-fields==1.0.2
drf-recaptcha==4.0.2
drf-spectacular==0.28.0
elastic-transport==9.1.0
elasticsearch==9.1.0
filelock==3.18.0
flake8==7.1.1
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.30.0
humanize==4.12.3
identify==2.6.12
idna==3.10
inflection==0.5.1
isort==5.12.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
lxml==6.0.0
Markdown==3.3.7
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
newsapi-python==0.2.7
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.0.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.2.0
PyMySQL==1.1.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ipware==3.0.0
python-twitter==3.5
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
slack_sdk==3.36.0
snscrape==0.7.0.20230622
soupsieve==2.7
sqlparse==0.5.3
stripe==12.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
vine==5.1.0
virtualenv==20.33.0
wcwidth==0.2.13
